+++ src/charm.py
@@ -158,6 +158,12 @@ class IngressConfiguratorCharm(ops.CharmBase):
 
     def _reconcile(self, _: ops.EventBase) -> None:
         """Dispatch to the appropriate reconcile method based on active relations."""
+        if self.app.planned_units() != 1:
+            self.unit.status = ops.BlockedStatus(
+                "ingress-configurator cannot have multiple units, scale down to a single unit"
+            )
+            return
+
         haproxy_route_related = self._haproxy_route.relation is not None
         haproxy_route_tcp_related = self._haproxy_route_tcp.relation is not None
         gateway_route_related = self._gateway_route.relation is not None